--- src/main/java/com/ahnsugyeong/dugout/exception/GeneralException.java
@@ -1,16 +1,0 @@
-package com.ahnsugyeong.dugout.exception;
-
-import com.ahnsugyeong.dugout.common.response.status.ErrorStatus;
-import lombok.Getter;
-
-@Getter
-public class GeneralException extends RuntimeException {
-
-    private final ErrorStatus errorStatus;
-
-    public GeneralException(ErrorStatus errorStatus) {
-        super(errorStatus.getMessage());
-        this.errorStatus = errorStatus;
-    }
-
-}
